requirements.txt
pandas>=2.2.3
tiktoken>=0.9.0
dashscope>=1.22.1
openai>=1.65.5 
chromadb>=0.5.23
PyYAML>=6.0